Move test setup into internal\setup
